Make a Bsseq object

Make a Bsseq object

Usage

Cpp_query_all(bedfiles, regions, bismark, merged, sparse, prealloc, nthreads)

Arguments

bedfiles

A vector of bedfiles

regions

A vector of regions

bismark

Whether the input is a bismark coverage file

prealloc

The number of rows to initialize the matrices with

nthreads

Set number of threads to use overriding the "iscream.threads" option. See ?set_threads for more information.