Make a BSSeq object using less memory

Create a BSSeq object using a encoded sparse methylation matrix. This method uses less memory than make_bsseq_hm since it uses sparse coverage and methylation matrices when creating the BSseq object. It is however slower in the BSseq object creation.

Usage

make_bsseq_lm(
  sparse_mat,
  sample_list,
  cov_filename = "cov.mtx.gz",
  m_filename = "M.mtx.gz"
)

Arguments

sparse_mat

A sparse matrix/long table of CpG loci and samples with beta and cov values

sample_list

A list of sample names for the bed files

cov_filename

Filename to write the coverage matrix market file to

m_filename

Filename to write the M matrix market file to

Value

A BSSeq object