Methylation matrix

Make a methylation matrix with given scWGBS sample bed files

Usage

make_meth_mat(
  sample_list,
  sample_path,
  cpg_bed_file,
  file_ext = ".bed.gz",
  merged = FALSE
)

Arguments

sample_list

A list of sample names for the bed files

sample_path

Path to the directory containing sample bed files

cpg_bed_file

The gzipped bed file with CpG loci. This is used to make the matrix index

file_ext

The extension of the cpg index and sample files

merged

Whether the bed files have CPs merged or unmerged